+++ params/types/genesisT/gen_genesis.go
@@ -30,7 +30,7 @@ import (
 	common0 "github.com/ethereum/go-ethereum/params/types/ctypes"
 	"github.com/ethereum/go-ethereum/params/types/goethereum"
 	"github.com/ethereum/go-ethereum/params/types/coregeth"
-	"github.com/ethereum/go-ethereum/params/types/multigethv0"
+	"github.com/ethereum/go-ethereum/params/types/multigeth"
 )
 
 var _ = (*genesisSpecMarshaling)(nil)
@@ -100,8 +100,8 @@ func (g *Genesis) UnmarshalJSON(input []byte) error {
 	switch conf.(type) {
 	case *coregeth.CoreGethChainConfig:
 		dec.Config = &coregeth.CoreGethChainConfig{}
-	case *multigethv0.ChainConfig:
-		dec.Config = &multigethv0.ChainConfig{}
+	case *multigeth.ChainConfig:
+		dec.Config = &multigeth.ChainConfig{}
 	case *goethereum.ChainConfig:
 		dec.Config = &goethereum.ChainConfig{}
 	default:
