+++ params/types/multigeth/doc.go
@@ -6,4 +6,4 @@ It is mostly used for managing upgrades between client versions,
 and is not encouraged for real use.
 */
 
-package multigethv0
+package multigeth
