+++ params/types/multigeth/multigethv0_chain_config_configurator.go
@@ -1,4 +1,4 @@
-package multigethv0
+package multigeth
 
 import (
 	"math/big"
